# Conflicts:
#	src/Core/tests/gtest_postgresql_protocol.cpp

+++ src/Core/PostgreSQLProtocol.h
@@ -1386,7 +1386,7 @@ private:
     String value;
 
 public:
-    CommandComplete(Command cmd_, Int32 rows_count_)
+    CommandComplete(Command cmd_, UInt64 rows_count_)
     {
         value = enum_to_string[cmd_];
